Move StreamGraph type out to a separate module
And stop importing VizGraph where it isn't strictly needed

This is necessary to avoid cycles in module imports later on.
