+++ src/Striot/CompileIoT.hs
@@ -1,9 +1,6 @@
 {-# OPTIONS_GHC -F -pgmF htfpp #-}
 
-module Striot.CompileIoT ( StreamGraph(..)
-                         , StreamVertex(..)
-                         , StreamOperator(..)
-                         , createPartitions
+module Striot.CompileIoT ( createPartitions
                          , generateCode
                          , GenerateOpts(..)
                          , defaultOpts
@@ -22,41 +19,7 @@ import Test.Framework
 import System.FilePath ((</>))
 import System.Directory (createDirectoryIfMissing)
 
-data StreamOperator = Map
-                    | Filter
-                    | Expand
-                    | Window
-                    | Merge
-                    | Join
-                    | Scan
-                    | FilterAcc
-                    | Source
-                    | Sink
-                    deriving (Ord,Eq)
-
-instance Show StreamOperator where
-    show Map             = "streamMap"
-    show Filter          = "streamFilter"
-    show Window          = "streamWindow"
-    show Merge           = "streamMerge"
-    show Join            = "streamJoin"
-    show Scan            = "streamScan"
-    show FilterAcc       = "streamFilterAcc"
-    show Expand          = "streamExpand"
-    show Source          = "streamSource"
-    show Sink            = "streamSink"
-
-
-data StreamVertex = StreamVertex
-    { vertexId   :: Int            -- Id needed for uniquely identifying a vertex. (Is there a nicer way?)
-    , operator   :: StreamOperator
-    , parameters :: [String]       -- operator arguments (excluding the input stream)
-    , intype     :: String
-    , outtype    :: String
-    } deriving (Eq,Show)
-
-instance Ord StreamVertex where
-    compare x y = compare (vertexId x) (vertexId y)
+import Striot.StreamGraph
 
 ------------------------------------------------------------------------------
 -- StreamGraph Partitioning
@@ -80,7 +43,6 @@ createPartitions g (p:ps) = ((overlay vs es):tailParts, cutEdges `overlay` tailC
 unPartition :: ([Graph StreamVertex], Graph StreamVertex) -> Graph StreamVertex
 unPartition (a,b) = overlay b $ foldl overlay Empty a
 
-type StreamGraph = Graph StreamVertex
 
 ------------------------------------------------------------------------------
 -- Code generation from StreamGraph definitions
@@ -283,31 +245,6 @@ s1 = path [ StreamVertex 0 (Source) [] "String" "String"
 test_reform_s0 = assertEqual s0 (unPartition $ createPartitions s0 [[0],[1]])
 test_reform_s1 = assertEqual s1 (unPartition $ createPartitions s1 [[0,1],[2]])
 test_reform_s1_2 = assertEqual s1 (unPartition $ createPartitions s1 [[0],[1,2]])
-
-------------------------------------------------------------------------------
--- quickcheck experiment
-
-instance Arbitrary StreamOperator where
-    arbitrary = elements [ Map , Filter , Expand , Window , Merge , Join , Scan
-                         , FilterAcc , Source , Sink ]
-
-instance Arbitrary StreamVertex where
-    arbitrary = do
-        vertexId <- arbitrary
-        operator <- arbitrary
-        let parameters = []
-            intype = "String"
-            outtype = "String"
-            in
-                return $ StreamVertex vertexId operator parameters intype outtype
-
-streamgraph :: Gen StreamGraph
-streamgraph = sized streamgraph'
-streamgraph' 0 = return g where g = empty :: StreamGraph
-streamgraph' n | n>0 = do
-    v <- arbitrary
-    t <- streamgraph' (n-1)
-    return $ connect (vertex v) t
 
 genDockerfile listen opts = 
     let pkgs = packages opts in concat
